# LATCH_GO_PORT_API_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Latch API",
    "version": "1.0.0",
    "description": "Machine-readable spec for porting Latch Next.js /api/* routes to Go. Extension-verified. See LATCH_GO_PORT_GUIDE.md."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Go dev server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Next.js reference server"
    }
  ],
...
  "paths": {
...
    "/api/multisig/drafts": {
      "post": {
        "summary": "POST /api/multisig/drafts",
        "operationId": "_api_multisig_drafts_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/route.ts",
        "x-latch-auth": "session",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL"
        ],
        "x-latch-db-models": [
          "MultisigDraft"
        ],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "get": {
        "summary": "GET /api/multisig/drafts",
        "operationId": "_api_multisig_drafts_get",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/route.ts",
        "x-latch-auth": "session",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL"
        ],
        "x-latch-db-models": [
          "MultisigDraft"
        ],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      }
    },
...
    "/api/multisig/join/{token}/members": {
      "post": {
        "summary": "POST /api/multisig/join/{token}/members",
        "operationId": "_api_multisig_join_token_members_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/join/[token]/members/route.ts",
        "x-latch-auth": "invite_token",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL"
        ],
        "x-latch-db-models": [
          "MultisigDraftMember"
        ],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/multisig/join/{token}/webauthn/register/finish": {
      "post": {
        "summary": "POST /api/multisig/join/{token}/webauthn/register/finish",
        "operationId": "_api_multisig_join_token_webauthn_register_finish_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/join/[token]/webauthn/register/finish/route.ts",
        "x-latch-auth": "invite_token",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL",
          "WEBAUTHN_RP_ID",
          "WEBAUTHN_ORIGIN"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/multisig/join/{token}/webauthn/authenticate/begin": {
      "post": {
        "summary": "POST /api/multisig/join/{token}/webauthn/authenticate/begin",
        "operationId": "_api_multisig_join_token_webauthn_authenticate_begin_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/join/[token]/webauthn/authenticate/begin/route.ts",
        "x-latch-auth": "invite_token",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL",
          "WEBAUTHN_RP_ID",
          "WEBAUTHN_ORIGIN"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid",
        "description": "30-day sliding session. Auto-created on first request."
      }
    },
    "schemas": {
...
      "LegacyErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}